feat(poc): Disable widget based on country
